+++ lib/src/db/models/attachment_model.dart
@@ -8,20 +8,37 @@ import '../../api/conversations/models/models.dart';
 class AttachmentModel extends Equatable {
   @Id()
   int? bid;
-  @Unique(onConflict: ConflictStrategy.replace)
+  @Unique()
   final String? fileId;
   final String? fileName;
   final String? fileBlurHash;
+  final String? url;
 
   AttachmentModel({
     this.bid,
     this.fileId,
     this.fileName,
     this.fileBlurHash,
+    this.url,
   });
 
+  AttachmentModel copyWith({
+    int? bid,
+    String? fileId,
+    String? fileName,
+    String? fileBlurHash,
+    String? url,
+  }) {
+    return AttachmentModel(
+        bid: bid ?? this.bid,
+        fileId: fileId ?? this.fileId,
+        fileName: fileName ?? this.fileName,
+        fileBlurHash: fileBlurHash ?? this.fileBlurHash,
+        url: url ?? this.url);
+  }
+
   @override
-  List<Object?> get props => [fileId, fileName, fileBlurHash];
+  List<Object?> get props => [fileId,fileName, fileBlurHash, url];
 }
 
 extension AttachmentModelExtension on Attachment {