+++ lib/src/features/conversation/bloc/conversation_bloc.dart
@@ -37,9 +37,9 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
   final MessagesRepository messagesRepository;
   final UserRepository userRepository;
 
-  StreamSubscription<ConversationModel>? updateConversationStreamSubscription;
   StreamSubscription<ChatMessage>? incomingMessagesSubscription;
   StreamSubscription<MessageSendStatus>? statusMessagesSubscription;
+  StreamSubscription<ConversationModel?>? conversationWatcher;
 
   ConversationBloc({
     required this.currentConversation,
@@ -48,7 +48,7 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
     required this.userRepository,
   }) : super(ConversationState(
             conversation: currentConversation,
-            participants: currentConversation.participants.toSet())) {
+            participants: Set.of(currentConversation.participants))) {
     on<MessagesRequested>(_onMessagesRequested);
     on<MessagesMoreRequested>(
       _onMessagesMoreRequested,
@@ -79,16 +79,6 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
 
     add(const ParticipantsReceived());
 
-    updateConversationStreamSubscription =
-        conversationRepository.updateConversationStream.listen((chat) async {
-      if (chat.id != currentConversation.id) return;
-
-      if (currentConversation != chat) {
-        currentConversation = currentConversation.copyWithItem(item: chat);
-        add(_ConversationUpdated(currentConversation));
-      }
-    });
-
     incomingMessagesSubscription =
         messagesRepository.incomingMessagesStream.listen((message) async {
       if (message.cid != currentConversation.id) return;
@@ -117,6 +107,15 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
           break;
       }
     });
+
+    conversationWatcher = messagesRepository.localDatasource
+        .watchedConversation(currentConversation.id)
+        .listen((chat) {
+      if (chat != null && chat != currentConversation) {
+        currentConversation = currentConversation.copyWithItem(item: chat);
+        add(_ConversationUpdated(currentConversation));
+      }
+    });
   }
 
   Future<void> _onMessagesRequested(
@@ -132,7 +131,7 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
               status: ConversationStatus.success,
               messages: messages,
               hasReachedMax: false,
-              participants: currentConversation.participants.toSet(),
+              participants: Set.of(currentConversation.participants),
               initial: true),
         );
         add(const MessagesRequested());
@@ -188,8 +187,8 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
 
   Future<void> _onParticipantsReceived(
       ParticipantsReceived event, Emitter<ConversationState> emit) async {
-    var participants =
-        await conversationRepository.updateParticipants(currentConversation);
+    var participants = await conversationRepository
+        .updateParticipants(currentConversation.copyWith());
     emit(state.copyWith(participants: Set.of(participants)));
   }
 
@@ -228,17 +227,17 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
       ),
     );
 
-    emit(state.copyWith(messages: messages));
+    emit(
+        state.copyWith(messages: messages, status: ConversationStatus.success));
   }
 
-  FutureOr<void> _onPendingStatusReceived(
-      _PendingStatusReceived event, Emitter<ConversationState> emit) {
+  Future<void> _onPendingStatusReceived(
+      _PendingStatusReceived event, Emitter<ConversationState> emit) async {
     var messages = [...state.messages];
 
     var msg = messages.firstWhere((o) => o.id == event.status.messageId);
-    messages[messages.indexOf(msg)] =
-        msg.copyWith(status: ChatMessageStatus.pending);
-
+    var msgUpdated = msg.copyWith(status: ChatMessageStatus.pending);
+    messages[messages.indexOf(msg)] = msgUpdated;
     emit(state.copyWith(messages: messages));
   }
 
@@ -261,7 +260,8 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
     var messages = {for (var v in state.messages) v.id!: v};
     var msgListUpdated = <MessageModel>[];
     event.status.msgIds?.forEach((id) {
-      if (messages[id]?.status != ChatMessageStatus.read) {
+      if (messages[id] != null &&
+          messages[id]?.status != ChatMessageStatus.read) {
         var msg = messages[id]!.copyWith(status: ChatMessageStatus.read);
         messages[id] = msg;
         msgListUpdated.add(msg);
@@ -276,9 +276,9 @@ class ConversationBloc extends Bloc<ConversationEvent, ConversationState> {
 
   @override
   Future<void> close() {
-    updateConversationStreamSubscription?.cancel();
     incomingMessagesSubscription?.cancel();
     statusMessagesSubscription?.cancel();
+    conversationWatcher?.cancel();
     return super.close();
   }
 }