+++ lib/src/navigation/app_router.dart
@@ -136,7 +136,7 @@ GoRouter router(BuildContext context, navigatorKey) => GoRouter(
         }
 
         if (status == AuthenticationStatus.authenticated) {
-          // fix for https://github.com/flutter/flutter/issues/146616
+          // fix for https://github.com/flutter/flutter/issues/146616 - ignoring Failed assertion for now
           if (state.fullPath ==
                   '$conversationListScreenPath/$conversationScreenSubPath' &&
               state.extra == null) {