+++ lib/src/shared/connection/bloc/connection_state.dart
@@ -3,14 +3,24 @@ part of 'connection_bloc.dart';
 enum ConnectionStatus { connecting, connected, disconnected, offline }
 
 final class ConnectionState extends Equatable {
-  const ConnectionState({this.status = ConnectionStatus.connecting});
+  const ConnectionState(
+      {this.status = ConnectionStatus.connecting,
+      AuthenticationStatus authStatus = AuthenticationStatus.unknown})
+      : _authStatus = authStatus;
 
   final ConnectionStatus status;
+  final AuthenticationStatus _authStatus;
 
-  ConnectionState copyWith({ConnectionStatus? status}) {
-    return ConnectionState(status: status ?? this.status);
+  ConnectionState copyWith({
+    ConnectionStatus? status,
+    AuthenticationStatus? authStatus,
+  }) {
+    return ConnectionState(
+      status: status ?? this.status,
+      authStatus: authStatus ?? _authStatus,
+    );
   }
 
   @override
-  List<Object> get props => [status];
+  List<Object> get props => [status, _authStatus];
 }